Change Mongoose to use wolfSSL instead of MBEDTLS
Configure Mongoose to use its OpenSSL adaptation layer and then set
the include path to get <openssl/ssl.h> from wolfSSL's OpenSSL
compatibility layer.  In mg_local.h we also need to include
user_settings.h from wolfssl so that the right conditionals are set
when processing the wolfSSL header files.  These conditionals turn on
the OpenSSL compatibility code in wolfSSL.

In wolfssl we need to add the source files from the main wolfssl layer
(before we only compiled the wolfcrypt sublayer).  We have to list all
the files to be compiled rather than taking the default action of
compiling all the sources in the SRCDIRS because the wolfssl tree
includes files containing assembly code for other architectures and I
want to leave the wolfssl tree intact.

To avoid a conflict between the definition of SHA_CTX in <openssl/ssl.h>
and in esp-idf/include/esp32/rom/sha.h we add #define SHA_CTX DUMMY
before reading the latter file to squelch its definition that we don't
need.